Make auto-join in ICB work by handling joins when not connected and updating the initial channel accordingly.
To make this work properly, move all initialization code from _didConnect
to _willConnect in the connections controller so that it is run before the
connection is made.  The chat protocols enqueue the commands and rerun them
in the right order once connected (doing any necessary processing, as is
the case of ICB).  Per timothy@'s suggestion.


git-svn-id: http://source.colloquy.info/svn/trunk@3594 cc480944-b4dd-0310-b5e3-89908df9b951